requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
auth==0.5.3
blinker==1.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
dnspython==1.16.0
ecdsa==0.17.0
eventlet==0.31.1
falcon==3.0.1
fastapi==0.68.0
Flask==2.0.1
greenlet==1.1.1
gunicorn==20.1.0
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mongoengine==0.23.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pymongo==3.12.0
pyOpenSSL==20.0.1
python-dotenv==0.19.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.23
starlette==0.14.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.15.0
Werkzeug==2.0.1
